Fix nomenclature, fix double play bug.
When choosing `spectator`, it would set the player name as
`playerspectator`. It's been changed to be just `spectator`.
Returning `true` from `.not` would double the `.val` listener, causing
it to play two tokens at a time (but only locally, meaning states would
diverge). Little API quirk that's been compensated for.