Switch to sass-loader, add game resets, code cleaning.
Instead of embedding the css rules in the html page, it's now being
managed by webpack and the react components that need it.
A game reset function has been added to reset game state. Since the game
list just loads pointers, CreateGame will think a game is safe and end
up loading an old game and nothing works and people are angry and hunt
me down with pitchforks. Instead, when flushing a game, it'll reset the
turn state and current occupants. Safe again!