Modularize container css, redesign active games page.
The container css is used on the homepage, and now the active games
page (I'm guessing I'll use it on the createGame page, too), so it's
been moved into it's own sass file and most sizing assumptions removed.
The GameList component has been redesigned to use the card design
component.