Redesign ChoosePlayer component, improve `disabled` feedback.
The screen where you choose a player, red, blue, or spectator, has been
completely redesigned. It uses the same responsive layout as the other 3
pages (Main, CreateGame, and GameList). It's also the most colorful, as
it uses the player color as it's background. It also shows when a player
is already taken by listing it as disabled, which brings to mind the
next point...
Disabled feedback has been improved. A consistent background is now
used, as well as setting the cursor to "not-allowed", and giving it some
consistently stylized background hover/active shading.