Mildly improved color scheme and positioning.
Player1 and player2 now have the new colors, and the status at the top
is centered. Also, the connect-four board is larger on mobile phones to
make things easier. Realistically though, I'm not sure how to make the
table look that good... it'll take a few more iterations before I'm as
proud of it as the rest of the app.