Add reuse of expired games.
When games are older than an hour, they're reset by the clients.
However, until recently, some data still hung around that would prevent
the object from being used again. Now, when it's purged, handlers are
put in place to make sure it's playable when a new game is created by
that name.