Squelch gun server messages, enable taken players.
Instead of blocking access to a player, simply mark it as taken. This
gives the user feedback without locking them out of a game (which would
happen if they refreshed).
More detail: you try to join a game, and it gives you three options:
player 1, player 2, and spectator. If you choose a player and then
refresh, the router hasn't had time to tell other people (including you)
that the player is now free. Since the buttons would be disabled if the
player was taken, as soon as you refresh, you're locked out of the game.
Rude.